updates
updates

+++ _output/msiafterburner/4.1.0/tools/chocolateyInstall.ps1
@@ -7,13 +7,14 @@ try {
     new-item "${env:temp}\afterburner" -itemtype directory
     $unpackfile = "${env:temp}\afterburner\afterburner.zip"
     $unpackdir = "${env:temp}\afterburner"
-    Stop-Process -ProcessName MSIAfterburner*
-    #$scriptPath = Split-Path -parent $MyInvocation.MyCommand.Definition
-    #$ahkFile = "$scriptPath\msiafterburnerinstall.ahk"
+    $ProcessActive = Get-Process MSIAfterburner* -ErrorAction SilentlyContinue
+    if($ProcessActive -ne $null)
+    {
+        Stop-Process -ProcessName MSIAfterburner*
+    }
     Get-ChocolateyWebFile $packageName $unpackfile $url
     Get-ChocolateyUnzip $unpackfile $unpackdir
     $file = "${env:temp}\afterburner\MSIAfterburnerSetup410.exe"
-    #Start-Process 'AutoHotkey' $ahkFile
     Install-ChocolateyInstallPackage $packageName $fileType $silentArgs $file 
     Remove-Item $unpackdir -Recurse -Force
     Write-ChocolateySuccess $packageName
